agregue una carpeta de estilos para las paginas
