+++ frontend/src/iconos.js
@@ -1,0 +1,44 @@
+// Archivo para centralizar los iconos usados en la aplicación
+
+import {
+  Car, 
+  Clock, 
+  Users, 
+  Settings, 
+  User, 
+  Phone, 
+  MapPin, 
+  Wrench, 
+  CheckCircle, 
+  PlayCircle, 
+  Edit,
+  Trash2,
+  AlertTriangle,
+  Flag,
+  Calendar,
+  TrendingUp,
+  Star,
+  Zap
+} from 'lucide-react';
+
+export {
+  Car, 
+  Clock, 
+  Users, 
+  Settings, 
+  User, 
+  Phone, 
+  MapPin, 
+  Wrench, 
+  CheckCircle, 
+  PlayCircle, 
+  Edit,
+  Trash2,
+  AlertTriangle,
+  Flag,
+  Calendar,
+  TrendingUp,
+  Star,
+  Zap
+};
+
