+++ frontend/src/main.jsx
@@ -4,7 +4,7 @@ import './index.css'
 import App from './App.jsx'
 import 'bootstrap/dist/css/bootstrap.min.css';
 
-
+// main es el punto de entrada de la aplicación
 createRoot(document.getElementById('root')).render(
   <StrictMode>
     <App />
