+++ frontend/src/styles/custom.css
@@ -19,19 +19,25 @@ body, html, #root {
 .hero-section {
     background: linear-gradient(90deg, #dc3545 0%, #c82333 100%);
     color: white;
-    padding: 4rem 0;
+    padding: 4rem 1rem;
     margin-top: 56px; /* Altura de la navbar */
     margin-bottom: auto;
     position: relative;
-    width: 100vw;
+    width: 100%;
+    max-width: 100vw;
 }
   
+.hero-section .text-center {
+    max-width: 1200px;         /* opcional: limita contenido centrado */
+    margin: 0 auto;
+}
+
 .hero-section::after {
     content: '';
     position: absolute;
     bottom: -40px;
     left: 0;
-    width: 100vw;
+    width: 100%;
     height: 100px;
     background: linear-gradient(90deg, #dc3545 0%, #c82333 100%);
     clip-path: ellipse(70% 100% at 50% 0%);
